refactor: 이미지 업로드 시 트랜잭션 범위 좁히도록 수정
